+++ src/main/java/hyve/petshow/repository/ServicoRepository.java
@@ -10,6 +10,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface ServicoRepository extends JpaRepository<Servico, Long> {
+public interface ServicoRepository extends JpaRepository<Servico, Integer> {
 	Optional<Servico> findByNome(String nome);
 }
